Remove BenchmarkGoalHeader and move its logic into [goal_id].tsx
