--- src/components/benchmarks/BenchmarkGoalHeader.tsx
@@ -1,94 +1,0 @@
-import { trpc } from "@/client/lib/trpc";
-import { useState } from "react";
-import Container from "@mui/material/Container";
-import $GoalPage from "@/styles/GoalPage.module.css";
-import { Grid } from "@mui/material";
-import $button from "@/components/design_system/button/Button.module.css";
-import { GoalHeader } from "../goal-header/goal-header";
-import { inferProcedureOutput } from "@trpc/server";
-import { AppRouter } from "@/backend/routers/_app";
-type GoalType = inferProcedureOutput<AppRouter["iep"]["getGoal"]>;
-const BenchmarkGoalHeader = ({
-  goal_index,
-  goal,
-}: {
-  goal_index: number;
-  goal: GoalType;
-}) => {
-  const [editGoal, setEditGoal] = useState(false);
-  const [editGoalInput, setEditGoalInput] = useState("");
-  const utils = trpc.useContext();
-  const showEditGoal = () => {
-    setEditGoal(true);
-    setEditGoalInput(goal?.description || "");
-  };
-
-  // TODO: modify callbacks for toast notification
-  const editMutation = trpc.iep.editGoal.useMutation({
-    onSuccess: () => utils.iep.getGoal.invalidate(),
-    onError: (err) => console.log({ err }),
-  });
-
-  const submitEditGoal = () => {
-    editMutation.mutate({
-      goal_id: goal.goal_id,
-      description: editGoalInput,
-    });
-    setEditGoal(false);
-    setEditGoalInput("");
-  };
-
-  const cancelEditGoal = () => {
-    setEditGoal(false);
-    setEditGoalInput("");
-  };
-
-  return (
-    <Container className={$GoalPage.goalDescriptionContainer}>
-      {!editGoal && (
-        <Grid container justifyContent="space-between">
-          {goal && goal_index && (
-            <GoalHeader
-              name={`Goal #${goal_index}`}
-              description={goal.description}
-              createdAt={goal.created_at}
-              goalId={goal.goal_id}
-              showEditGoal={showEditGoal}
-            />
-          )}
-        </Grid>
-      )}
-      {editGoal && (
-        <form style={{ marginTop: "1rem" }} onSubmit={submitEditGoal}>
-          <textarea
-            value={editGoalInput}
-            name="description"
-            onChange={(e) => {
-              setEditGoalInput(e.target.value);
-            }}
-            className={$GoalPage.editGoalFormTextArea}
-          />
-          <Grid
-            container
-            justifyContent="space-around"
-            marginY={2}
-            md={6}
-            marginX="auto"
-          >
-            <Grid item>
-              <button className={$button.secondary} onClick={cancelEditGoal}>
-                Cancel
-              </button>
-            </Grid>
-            <Grid item>
-              <button className={$button.default} type="submit">
-                Save
-              </button>
-            </Grid>
-          </Grid>
-        </form>
-      )}
-    </Container>
-  );
-};
-export default BenchmarkGoalHeader;
