+++ src/pages/students/[student_id]/goals/[goal_id].tsx
@@ -3,12 +3,20 @@ import { useRouter } from "next/router";
 import Stack from "@mui/material/Stack";
 import Grid from "@mui/material/Grid";
 import BenchmarksContainer from "@/components/benchmarks/BenchmarksContainer";
-import BenchmarkGoalHeader from "@/components/benchmarks/BenchmarkGoalHeader";
+import { GoalHeader } from "@/components/goal-header/goal-header";
+import { useState } from "react";
+import $GoalPage from "@/styles/GoalPage.module.css";
+import { Container } from "@mui/material";
+import $button from "@/components/design_system/button/Button.module.css";
 
 const GoalPage = () => {
   const router = useRouter();
   const goal_id = (router.query?.goal_id as string) || "";
   const student_id = (router.query?.student_id as string) || "";
+  const [editGoal, setEditGoal] = useState(false);
+  const [editGoalInput, setEditGoalInput] = useState("");
+  const utils = trpc.useContext();
+
   const { data: activeIep } = trpc.student.getActiveStudentIep.useQuery(
     { student_id: student_id },
     { enabled: Boolean(student_id), retry: false }
@@ -26,8 +34,32 @@ const GoalPage = () => {
     { goal_id: goal_id },
     { enabled: Boolean(goal_id) }
   );
-  const goal_index =
-    goals.findIndex((g) => g.goal_id === goal?.goal_id) + 1 || 0;
+
+  const showEditGoal = () => {
+    setEditGoal(true);
+    setEditGoalInput(goal?.description || "");
+  };
+
+  // TODO: modify callbacks for toast notification
+  const editMutation = trpc.iep.editGoal.useMutation({
+    onSuccess: () => utils.iep.getGoal.invalidate(),
+    onError: (err) => console.log({ err }),
+  });
+
+  const submitEditGoal = () => {
+    editMutation.mutate({
+      goal_id: goal.goal_id,
+      description: editGoalInput,
+    });
+    setEditGoal(false);
+    setEditGoalInput("");
+  };
+
+  const cancelEditGoal = () => {
+    setEditGoal(false);
+    setEditGoalInput("");
+  };
+  const goal_index = goals.findIndex((g) => g.goal_id === goal?.goal_id) + 1;
   return (
     <Stack
       spacing={2}
@@ -41,7 +73,51 @@ const GoalPage = () => {
       }}
     >
       {/* Goal Description */}
-      {goal && <BenchmarkGoalHeader goal_index={goal_index} goal={goal} />}
+      <Container className={$GoalPage.goalDescriptionContainer}>
+        {!editGoal && (
+          <Grid container justifyContent="space-between">
+            {goal && goal_index && (
+              <GoalHeader
+                name={`Goal #${goal_index}`}
+                description={goal.description}
+                createdAt={goal.created_at}
+                goalId={goal.goal_id}
+                showEditGoal={showEditGoal}
+              />
+            )}
+          </Grid>
+        )}
+        {editGoal && (
+          <form style={{ marginTop: "1rem" }} onSubmit={submitEditGoal}>
+            <textarea
+              value={editGoalInput}
+              name="description"
+              onChange={(e) => {
+                setEditGoalInput(e.target.value);
+              }}
+              className={$GoalPage.editGoalFormTextArea}
+            />
+            <Grid
+              container
+              justifyContent="space-around"
+              marginY={2}
+              md={6}
+              marginX="auto"
+            >
+              <Grid item>
+                <button className={$button.secondary} onClick={cancelEditGoal}>
+                  Cancel
+                </button>
+              </Grid>
+              <Grid item>
+                <button className={$button.default} type="submit">
+                  Save
+                </button>
+              </Grid>
+            </Grid>
+          </form>
+        )}
+      </Container>
       {/* Benchmarks */}
       <Grid container justifyContent="space-between">
         <Grid item>
